LPS-56703 Move StagedModelDataHandlers to OSGi module
