+++ modules/apps/document-library/document-library-service/src/com/liferay/document/library/lar/xstream/FileEntryConverter.java
@@ -12,13 +12,15 @@
  * details.
  */
 
-package com.liferay.portlet.documentlibrary.lar.xstream;
+package com.liferay.document.library.lar.xstream;
 
 import com.liferay.portal.kernel.bean.BeanPropertiesUtil;
-import com.liferay.portal.repository.liferayrepository.model.LiferayFolder;
-import com.liferay.portal.repository.proxy.FolderProxyBean;
-import com.liferay.portlet.documentlibrary.model.DLFolder;
-import com.liferay.portlet.documentlibrary.model.impl.DLFolderImpl;
+import com.liferay.portal.repository.liferayrepository.model.LiferayFileEntry;
+import com.liferay.portal.repository.liferayrepository.model.LiferayFileVersion;
+import com.liferay.portal.repository.proxy.FileEntryProxyBean;
+import com.liferay.portal.repository.proxy.FileVersionProxyBean;
+import com.liferay.portlet.documentlibrary.model.DLFileEntry;
+import com.liferay.portlet.documentlibrary.model.impl.DLFileEntryImpl;
 import com.liferay.portlet.exportimport.xstream.BaseXStreamConverter;
 import com.liferay.portlet.exportimport.xstream.XStreamHierarchicalStreamReader;
 import com.liferay.portlet.exportimport.xstream.XStreamUnmarshallingContext;
@@ -29,11 +31,11 @@ import java.util.List;
 /**
  * @author Akos Thurzo
  */
-public class FolderConverter extends BaseXStreamConverter {
+public class FileEntryConverter extends BaseXStreamConverter {
 
 	@Override
 	public boolean canConvert(Class<?> clazz) {
-		return clazz.equals(FolderProxyBean.class);
+		return clazz.equals(FileEntryProxyBean.class);
 	}
 
 	@Override
@@ -42,8 +44,9 @@ public class FolderConverter extends BaseXStreamConverter {
 			XStreamUnmarshallingContext xStreamUnmarshallingContext)
 		throws Exception {
 
-		DLFolder dlFolder = new DLFolderImpl();
+		DLFileEntry dlFileEntry = new DLFileEntryImpl();
 		boolean escapedModel = false;
+		LiferayFileVersion liferayFileVersion = null;
 
 		while (xStreamHierarchicalStreamReader.hasMoreChildren()) {
 			xStreamHierarchicalStreamReader.moveDown();
@@ -51,7 +54,11 @@ public class FolderConverter extends BaseXStreamConverter {
 			String nodeName = xStreamHierarchicalStreamReader.getNodeName();
 
 			Class<?> clazz = BeanPropertiesUtil.getObjectType(
-				dlFolder, nodeName);
+				dlFileEntry, nodeName);
+
+			if (nodeName.equals(FieldConstants.FILE_VERSION)) {
+				clazz = FileVersionProxyBean.class;
+			}
 
 			Object convertedValue = xStreamUnmarshallingContext.convertAnother(
 				xStreamHierarchicalStreamReader.getValue(), clazz);
@@ -60,16 +67,24 @@ public class FolderConverter extends BaseXStreamConverter {
 				if (nodeName.equals(FieldConstants.ESCAPED_MODEL)) {
 					escapedModel = (Boolean)convertedValue;
 				}
+				else if (nodeName.equals(FieldConstants.FILE_VERSION)) {
+					liferayFileVersion = (LiferayFileVersion)convertedValue;
+				}
 				else {
 					BeanPropertiesUtil.setProperty(
-						dlFolder, nodeName, convertedValue);
+						dlFileEntry, nodeName, convertedValue);
 				}
 			}
 
 			xStreamHierarchicalStreamReader.moveUp();
 		}
 
-		return new LiferayFolder(dlFolder, escapedModel);
+		LiferayFileEntry liferayFileEntry = new LiferayFileEntry(
+			dlFileEntry, escapedModel);
+
+		liferayFileEntry.setCachedFileVersion(liferayFileVersion);
+
+		return liferayFileEntry;
 	}
 
 	@Override
@@ -84,18 +99,23 @@ public class FolderConverter extends BaseXStreamConverter {
 		fields.add(FieldConstants.CREATE_DATE);
 		fields.add(FieldConstants.DESCRIPTION);
 		fields.add(FieldConstants.ESCAPED_MODEL);
+		fields.add(FieldConstants.EXTENSION);
+		fields.add(FieldConstants.FILE_ENTRY_ID);
+		fields.add(FieldConstants.FILE_VERSION);
 		fields.add(FieldConstants.FOLDER_ID);
 		fields.add(FieldConstants.GROUP_ID);
-		fields.add(FieldConstants.LAST_POST_DATE);
+		fields.add(FieldConstants.MANUAL_CHECK_IN_REQUIRED);
+		fields.add(FieldConstants.MIME_TYPE);
 		fields.add(FieldConstants.MODIFIED_DATE);
-		fields.add(FieldConstants.MOUNT_POINT);
-		fields.add(FieldConstants.NAME);
-		fields.add(FieldConstants.PARENT_FOLDER_ID);
+		fields.add(FieldConstants.READ_COUNT);
 		fields.add(FieldConstants.REPOSITORY_ID);
+		fields.add(FieldConstants.SIZE);
+		fields.add(FieldConstants.TITLE);
 		fields.add(FieldConstants.USER_ID);
 		fields.add(FieldConstants.USER_NAME);
 		fields.add(FieldConstants.USER_UUID);
 		fields.add(FieldConstants.UUID);
+		fields.add(FieldConstants.VERSION);
 	}
 
 }
